+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.0 (unreleased)
+## 0.2.0 (2021-10-17)
 
 - Added support for Windows